Enabled Firebase authentication and Firestore
I completed the backend for the code, the users phone number, house address and email will now be saved to Firestore. The user is now able to be authenticated by firebase authentication. In the event the user forgets their password they can click the forgot password section and a recovery email will be sent.